refactor: changed data type 'int' to 'long' in Id property

+++ WAW.API/Company/Domain/Repositories/ICompanyRepository.cs
@@ -5,7 +5,7 @@ using WAW.API.Company.Domain.Models;
 public interface ICompanyRepository {
   Task<IEnumerable<Company>> ListAll();
   Task Add(Company company);
-  Task<Company?> FindById(int id);
+  Task<Company?> FindById(long id);
   Task<Company?> FindByName(string name);
   void Update(Company company);
   void Remove(Company company);
