+++ WAW.API/Company/Domain/Services/ICompanyService.cs
@@ -7,6 +7,6 @@ using WAW.API.Company.Domain.Models;
 public interface ICompanyService {
   Task<IEnumerable<Company>> ListAll();
   Task<CompanyResponse> Create(Company company);
-  Task<CompanyResponse> Update(int id, Company company);
-  Task<CompanyResponse> Delete(int id);
+  Task<CompanyResponse> Update(long id, Company company);
+  Task<CompanyResponse> Delete(long id);
 }
